focusfollowmouse: give precedence to the mouse cursor when deciding next focus
By default dwm will revert focus back to the visible client that last had focus
when moving windows into and out of view.

This can happen when the user:
   - changes tags
   - toggles tags into or out of view
   - moves the selected window to another monitor
   - moves the selected window to another tag
   - a window closes

With this patch the expectation is that the window that remains under the mouse
cursor will be the next one to receive focus.